float -> double & fix warnings

+++ src/Constants.h
@@ -8,26 +8,25 @@
 #ifndef PidConstants_H
 #define PidConstants_H 1
 
-namespace PidParticles{
-    enum eParticle{
-        kProton = 0,
-        kPion,
-        kKaon,
-        kBg,
-        nParticles
-    };
-    
-    const float masses[nParticles] = { 0.938, 0.140, 0.5, 0.0 };
+namespace PidParticles {
+enum eParticle {
+  kProton = 0,
+  kPion,
+  kKaon,
+  kBg,
+  nParticles
+};
+
+const double masses[nParticles] = {0.938, 0.140, 0.5, 0.0};
 }
 
-namespace PidFunction{
-    enum eNames{
-        kA = 0,
-        kMean,
-        kSigma,
-        nParams
-    };    
+namespace PidFunction {
+enum eNames {
+  kA = 0,
+  kMean,
+  kSigma,
+  nParams
+};
 }
-
 
 #endif
